dashboard/pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = ""
authors = ["niklas <niklas.laufkoetter@nateurope.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
dash = "^2.6.2"
flask = "^2.2.2"
dash-bootstrap-components = "^1.2.1"
python-decouple = "^3.6"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
